Remove CoC and Security files.
The Code of Conduct file is managed through the special `.github` repository. This allows the organization to update the file once for all repositories.

If there is a compelling reason as to why a certain repository should have its own `CODE_OF_CONDUCT.md` file, all changes to the file should be submitted through a pull request to the WordPress/.github repository.

Additionally, an empty `SECURITY.md` file is not helpful and blocks the security team from adding a default file to the `.github` repo in the future (which is currently being explored).